refactor: move spellcasting directories
