+++ src/core/models/CMakeLists.txt
@@ -12,3 +12,4 @@ add_subdirectory(effects)
 add_subdirectory(feature)
 add_subdirectory(item)
 add_subdirectory(spell)
+add_subdirectory(spellcasting)
