+++ src/core/models/character_class/character_class.cpp
@@ -18,7 +18,7 @@
 #include <core/errors/validation_error.hpp>
 #include <core/exceptions/validation_exceptions.hpp>
 #include <core/models/character_class/important_levels.hpp>
-#include <core/models/character_class/spellcasting/spellcasting_factory.hpp>
+#include <core/models/spellcasting/spellcasting_factory.hpp>
 #include <core/models/feature/class_feature.hpp>
 #include <core/models/source_info.hpp>
 #include <core/validation/character_class/character_class_data.hpp>
