+++ src/core/models/character_subclass/character_subclass.cpp
@@ -11,7 +11,7 @@
 #include <core/content.hpp>
 #include <core/errors/errors.hpp>
 #include <core/exceptions/validation_exceptions.hpp>
-#include <core/models/character_class/spellcasting/spellcasting_factory.hpp>
+#include <core/models/spellcasting/spellcasting_factory.hpp>
 #include <core/models/character_race/character_race.hpp>
 #include <core/models/feature/class_feature.hpp>
 #include <core/models/source_info.hpp>
