+++ src/core/models/character_subclass/character_subclass.hpp
@@ -9,7 +9,7 @@
 #include <vector>
 
 #include <core/models/character_class/character_class.hpp>
-#include <core/models/character_class/spellcasting/spellcasting.hpp>
+#include <core/models/spellcasting/spellcasting.hpp>
 #include <core/models/content_piece.hpp>
 #include <core/models/feature/class_feature.hpp>
 #include <core/models/source_info.hpp>
