+++ src/core/models/spellcasting/CMakeLists.txt
