+++ src/core/models/spellcasting/preparation_spellcasting.cpp
@@ -8,7 +8,7 @@
 
 #include <core/basic_mechanics/abilities.hpp>
 #include <core/models/character/ability_scores.hpp>
-#include <core/models/character_class/spellcasting/spellcasting.hpp>
+#include <core/models/spellcasting/spellcasting.hpp>
 
 dnd::PreparationSpellcastingType dnd::PreparationSpellcasting::get_type() const noexcept { return type; }
