+++ src/core/models/spellcasting/spellcasting.cpp
