+++ src/core/models/spellcasting/spellcasting.hpp
