+++ src/core/models/spellcasting/spells_known_spellcasting.hpp
@@ -6,7 +6,7 @@
 #include <array>
 
 #include <core/basic_mechanics/abilities.hpp>
-#include <core/models/character_class/spellcasting/spellcasting.hpp>
+#include <core/models/spellcasting/spellcasting.hpp>
 
 namespace dnd {
