+++ src/core/validation/CMakeLists.txt
@@ -14,3 +14,4 @@ add_subdirectory(effects)
 add_subdirectory(feature)
 add_subdirectory(item)
 add_subdirectory(spell)
+add_subdirectory(spellcasting)
