+++ src/core/validation/character_class/CMakeLists.txt
@@ -3,5 +3,3 @@ target_sources(${DND_CORE}
     character_class_data.cpp
     important_levels_data.cpp
 )
-
-add_subdirectory(spellcasting)
