+++ src/core/validation/spellcasting/CMakeLists.txt
