+++ src/core/validation/spellcasting/spellcasting_data.cpp
