+++ tests/testcore/validation/character_class/CMakeLists.txt
@@ -3,5 +3,3 @@ target_sources(${DND_TESTS}
     character_class_data_test.cpp
     important_levels_data_test.cpp
 )
-
-add_subdirectory(spellcasting)
