+++ tests/testcore/validation/spellcasting/CMakeLists.txt
